third_party/WebKit/LayoutTests/fast/events/clipboard-events-synthetic.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<script>
// 'before*' events are non-standard.
// For standard events ('copy', 'cut', 'paste') are covered by web platform
// tests.
const EVENTS = [ 'beforecopy', 'beforecut', 'beforepaste' ];

EVENTS.forEach(name => {
  function testEvent(name, init, composed_flag_expectation, name) {
    async_test(function(test) {
      document.addEventListener(name, test.step_func_done((e) => {
        assert_false(e.isTrusted, `synthetic ${name} event is untrusted`);
        assert_equals(e.composed, composed_flag_expectation,
                      `composed flag should be ${composed_flag_expectation}`);
      }));
      let event = new ClipboardEvent(name, init);
      document.dispatchEvent(event);
    }, name);
  }

  testEvent(name, { bubbles: true, cancellable: true}, false,
            `Unspecified synthetic ${name} event should not be composed.`);
  testEvent(name, { bubbles: true, cancelable: true, composed: true }, true,
            `Synthetic ${name} event can be explicitly composed.`);
  testEvent(name, { bubbles: true, cancelable: true, composed: false }, false,
            `Synthetic ${name} event can be explicitly uncomposed.`);
});
</script>
